Create Selection Sort.cpp
Selection Sort works by dividing the array into a sorted and an unsorted region. 

In each iteration, it finds the minimum element from the unsorted region and swaps it with the first element of the unsorted region. 

This process is repeated until the entire array is sorted.

Only one swap in one iteration